Flow legend, flow density options, fixed various display bugs

+++ scripts/Flox.js
@@ -21,13 +21,83 @@ var mapComponent,
 	},
 	startTimeAll, 
 	endTimeAll,
+	fipsLookupTable,
 	my = {};
+
+fipsLookupTable = {
+	"01": "Alabama",//AL
+	"02": "Alaska",	//AK
+	"04": "Arizona",//AZ
+	"05": "Arkansas",//AR
+	"06": "California",//CA
+	"08": "Colorado",//CO
+	"09": "Connecticut",//CT
+	"10": "Delaware",//DE
+	"11": "District of Columbia",//DC
+	"12": "Florida",//FL
+	"13": "Georgia",//GA
+	"15": "Hawaii",	//HI
+	"16": "Idaho",//ID
+	"17": "Illinois",//IL
+	"18": "Indiana",//IN
+	"19": "Iowa",//IA
+	"20": "Kansas",//KS
+	"21": "Kentucky",//KY
+	"22": "Louisiana",//LA
+	"23": "Maine",//ME
+	"24": "Maryland",//MD
+	"25": "Massachusetts",//MA
+	"26": "Michigan",//MI
+	"27": "Minnesota",//MN
+	"28": "Mississippi",//MS
+	"29": "Missouri",//MO
+	"30": "Montana"	,//MT
+	"31": "Nebraska",//NE
+	"32": "Nevada",//NV
+	"33": "New Hampshire",//NH
+	"34": "New Jersey",//NJ
+	"35": "New Mexico",//NM
+	"36": "New York",//NY
+	"37": "North Carolina",	//NC
+	"38": "North Dakota",//ND
+	"39": "Ohio",//OH
+	"40": "Oklahoma",	//OK
+	"41": "Oregon",	//OR
+	"42": "Pennsylvania",//PA
+	"44": "Rhode Island",//RI
+	"45": "South Carolina",//SC
+	"46": "South Dakota",//SD
+	"47": "Tennessee",//TN
+	"48": "Texas",//TX
+	"49": "Utah",//UT
+	"50": "Vermont",//VT
+	"51": "Virginia",//VA
+	"53": "Washington",//WA
+	"54": "West Virginia",//WV
+	"55": "Wisconsin",//WI
+	"56": "Wyoming",//WY
+	"60": "American Samoa",//AS
+	"64": "Federated States of Micronesia",//FM
+	"66": "Guam",//GU	1
+	"68": "Marshall Islands",//MH	3
+	"69": "Commonwealth of the Northern Mariana Islands",//MP
+	"70": "Palau",//PW
+	"72": "Puerto Rico",//PR
+	"74": "U.S. Minor Outlying Islands",//UM
+	"78": "U.S. Virgin Islands"//VI
+}
 
 function refreshMap(model_copy) {
 	if(!model_copy) {
 		throw new Error("refreshMap needs a model passed in");
 	}
     mapComponent.drawFeatures(model_copy);
+}
+
+function terminateWorkers() {
+	if (importWorker) {importWorker.terminate();}
+	if (filterWorker) {filterWorker.terminate();}
+	if (layoutWorker) {layoutWorker.terminate();}
 }
 
 // TODO If the browser can't do webworkers, then webworkers shouldn't be used.
@@ -41,17 +111,21 @@ function initLayoutWorker(modelCopy, callback) {
 		// If a layouter worker currently exists, call terminate on it. 
 		// If it was in the middle of something, it'll stop and do this
 		// instead. If it wasn't doing anything, then this won't matter.
-		if(layoutWorker) {
-			layoutWorker.terminate();
-		}
+		if(layoutWorker) {layoutWorker.terminate();}
+		
 		// Web workers take a separate file. Note that the path is relative
 		// to index.html, not Flox.js
 		layoutWorker = new Worker("scripts/layoutWorker.js");
 		
 		// This happens when layoutWorker sends out a message
 		layoutWorker.onmessage = function(e) {
+			var progress;
+			
+			// progress is 50 plus 
+			
+			
 			// Update the progress bar.
-			Flox.GUI.updateLayoutProgressBar(e.data[1]);
+			Flox.GUI.updateLayoutProgressBar(50 + (e.data[1]/2));
 			
 			// Get the new control points
 			ctrlPts = e.data[0];
@@ -117,13 +191,18 @@ function initImportWorker(callback) {
 		if (importWorker) {importWorker.terminate();}
 		importWorker = new Worker("importWorker.js");
 		importWorker.onmessage = function(e) {	
-			// Send the imported data to whoever called runImportWorker		
+			// update progress bar
+			Flox.GUI.updateLayoutProgressBar(30);
+			
+			// Send the imported data to whoever called runImportWorker
 			callback(e.data);
 		}
 	}	
 }
 
 function runImportWorker(stuffImportWorkerNeeds, callback) {
+	Flox.GUI.showLayoutProgressBar();
+	Flox.GUI.updateLayoutProgressBar(10);
 	initImportWorker(callback);
 
 	// post a message to the import worker
@@ -135,6 +214,7 @@ function runImportWorker(stuffImportWorkerNeeds, callback) {
  * with the appropriate weight.
  * Decreases the weight of each iteration. 
  * Calls moveFlowsIntersectingNodes() during second half of iterations.
+ * NOTE: this is run only if webworkers are not enabled or available
  */
 function layoutFlows(model) {
 	
@@ -185,6 +265,7 @@ function layoutFlows(model) {
 
 function importStateToStateMigrationFlows(keepSelectedState) {
 	// clear the model
+	terminateWorkers();
     model_master.deleteAllFlows();
     mapComponent.hideAllCountyBorders();
     mapComponent.removeAllFlows();
@@ -361,6 +442,7 @@ my.importStateToStateMigrationFlows = function (keepSelectedState) {
 };
 
 my.importTotalCountyFlowData = function(stateFIPS) {		
+	terminateWorkers();
 	
 	var stuffImportWorkerNeeds = {};
 	
@@ -424,9 +506,12 @@ my.filterBySettings = function(m, settings) {
  * Filters, performs layout, and displays the model according to current settings.
  */
 my.updateMap = function() {
+	// There should be no workers running right now.
+	terminateWorkers();
+	mapComponent.clearAllMapFeatures();
 	
-	Flox.GUI.updateLayoutProgressBar(0);
 	Flox.GUI.showLayoutProgressBar();
+	
 	// Good time to assign xy coordinates to nodes.
 	my.assignXYToNodes(model_master.getPoints());
 
@@ -437,13 +522,16 @@ my.updateMap = function() {
 	// FIXME Something is broken when not using webworkers. 
 	Flox.GUI.updateGUI();
 	if(window.Worker && model_master.settings.useWebworkers) {
+		Flox.GUI.updateLayoutProgressBar(35);
 		runFilterWorker(function(filteredModel) {
+			Flox.GUI.updateLayoutProgressBar(50);
 			// configure the needed variables to get only above average flows.
 			filteredModel.setAboveAverageFlowCount();
 			if(filterSettings.stateMode === false &&
 				filterSettings.selectedState !== false) {
 				mapComponent.configureNecklaceMap(filteredModel);
 			}
+			mapComponent.enableTooltip();
 			runLayoutWorker(filteredModel, function() {
 				
 				refreshMap(filteredModel);
@@ -678,6 +766,38 @@ my.getNodeCoordinates = function() {
 
 my.getPopulationDensityColor = function() {
 	return mapComponent.getPopulationDensityColor();
+};
+
+my.enableTooltip = function() {
+	mapComponent.enableTooltip();
+};
+
+my.disableTooltip = function() {
+	mapComponent.disableTooltip();
+};
+
+// Get a FIPS, return a state name.
+my.lookupFIPS = function(FIPS) {
+	// if the fips is a number, make it at string. Or just string it anyway.
+	FIPS = String(FIPS);
+	
+	// If length of FIPS is 1, then it's a single-digit number. Add leading 0.
+	if(FIPS.length === 1) {
+		FIPS = "0" + FIPS;
+	}
+	
+	if(!fipsLookupTable.hasOwnProperty(FIPS)) {
+		throw new Error(FIPS + " is not a legit FIPS code");
+	} 
+
+	return fipsLookupTable[FIPS];
+};
+
+my.getSelectedStateName = function() {
+	if(filterSettings.selectedState === false) {
+		throw new Error("No state is selected");
+	}
+	return my.lookupFIPS(filterSettings.selectedState);
 };
 
 // END DEBUG STUFF-------------------------------------
